requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
discord.py==2.0.1
distlib==0.3.6
filelock==3.8.0
frozenlist==1.3.1
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
multidict==6.0.2
platformdirs==2.5.2
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.4
yarl==1.8.1